New file structure for v2.0.0 (#248)
