+++ minigrid/envs/crossing.py
@@ -2,7 +2,10 @@ import itertools as itt
 
 import numpy as np
 
-from minigrid.minigrid import Goal, Grid, Lava, MiniGridEnv, MissionSpace
+from minigrid.core.grid import Grid
+from minigrid.core.mission import MissionSpace
+from minigrid.core.world_object import Goal, Lava
+from minigrid.minigrid import MiniGridEnv
 
 
 class CrossingEnv(MiniGridEnv):
