+++ minigrid/envs/distshift.py
@@ -1,4 +1,7 @@
-from minigrid.minigrid import Goal, Grid, Lava, MiniGridEnv, MissionSpace
+from minigrid.core.grid import Grid
+from minigrid.core.mission import MissionSpace
+from minigrid.core.world_object import Goal, Lava
+from minigrid.minigrid import MiniGridEnv
 
 
 class DistShiftEnv(MiniGridEnv):
