+++ minigrid/envs/empty.py
@@ -1,8 +1,10 @@
-from minigrid.minigrid import Goal, Grid, MiniGridEnv, MissionSpace
+from minigrid.core.grid import Grid
+from minigrid.core.mission import MissionSpace
+from minigrid.core.world_object import Goal
+from minigrid.minigrid import MiniGridEnv
 
 
 class EmptyEnv(MiniGridEnv):
-
     """
     ### Description
