+++ minigrid/envs/fetch.py
@@ -1,4 +1,8 @@
-from minigrid.minigrid import COLOR_NAMES, Ball, Grid, Key, MiniGridEnv, MissionSpace
+from minigrid.core.constants import COLOR_NAMES
+from minigrid.core.grid import Grid
+from minigrid.core.mission import MissionSpace
+from minigrid.core.world_object import Ball, Key
+from minigrid.minigrid import MiniGridEnv
 
 
 class FetchEnv(MiniGridEnv):
