+++ minigrid/envs/multiroom.py
@@ -1,12 +1,8 @@
-from minigrid.minigrid import (
-    COLOR_NAMES,
-    Door,
-    Goal,
-    Grid,
-    MiniGridEnv,
-    MissionSpace,
-    Wall,
-)
+from minigrid.core.constants import COLOR_NAMES
+from minigrid.core.grid import Grid
+from minigrid.core.mission import MissionSpace
+from minigrid.core.world_object import Door, Goal, Wall
+from minigrid.minigrid import MiniGridEnv
 
 
 class MultiRoom:
