+++ minigrid/envs/playground.py
@@ -1,13 +1,8 @@
-from minigrid.minigrid import (
-    COLOR_NAMES,
-    Ball,
-    Box,
-    Door,
-    Grid,
-    Key,
-    MiniGridEnv,
-    MissionSpace,
-)
+from minigrid.core.constants import COLOR_NAMES
+from minigrid.core.grid import Grid
+from minigrid.core.mission import MissionSpace
+from minigrid.core.world_object import Ball, Box, Door, Key
+from minigrid.minigrid import MiniGridEnv
 
 
 class PlaygroundEnv(MiniGridEnv):
