+++ minigrid/utils/window.py
