+++ minigrid/wrappers.py
@@ -7,7 +7,8 @@ import numpy as np
 from gymnasium import spaces
 from gymnasium.core import ObservationWrapper, Wrapper
 
-from minigrid.minigrid import COLOR_TO_IDX, OBJECT_TO_IDX, STATE_TO_IDX, Goal
+from minigrid.core.constants import COLOR_TO_IDX, OBJECT_TO_IDX, STATE_TO_IDX
+from minigrid.core.world_object import Goal
 
 
 class ReseedWrapper(Wrapper):
