+++ tests/test_envs.py
@@ -6,7 +6,8 @@ import pytest
 from gymnasium.envs.registration import EnvSpec
 from gymnasium.utils.env_checker import check_env
 
-from minigrid.minigrid import Grid, MissionSpace
+from minigrid.core.grid import Grid
+from minigrid.core.mission import MissionSpace
 from tests.utils import all_testing_env_specs, assert_equals
 
 CHECK_ENV_IGNORE_WARNINGS = [
